requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
pyodbc

# Azure services
azure-identity
azure-storage-blob
azure-keyvault-secrets

# Data validation
pydantic

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OAuth support
requests-oauthlib==1.3.1
oauthlib==3.2.2

# HTTP requests
httpx
requests

# Utility libraries
python-dotenv
python-dateutil==2.9.0.post0
jinja2

# Dev & testing
pytest
httpx

# Optional: for async SQL Server support
# sqlalchemy[asyncio]
# aiosqlite